# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RBAC Authorization Platform API
  description: Multi-tenant authorization platform implementing RBAC, ABAC, and policy-based access control
  version: 1.0.0
  contact:
    name: Authorization Platform Team
    email: team@example.com

servers:
  - url: http://localhost:3000
    description: Authorization Engine (Development)
  - url: http://localhost:3001
    description: Management API (Development)
  - url: http://localhost:3002
    description: Audit Service (Development)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - is_active

    Principal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        is_active:
          type: boolean
        attributes:
          type: object
        created_at:
          type: string
          format: date-time
      required:
        - id
        - tenant_id
        - username
        - is_active

    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        is_system:
          type: boolean
        permission_ids:
          type: array
          items:
            type: string
            format: uuid
        created_at:
          type: string
          format: date-time
      required:
        - id
        - tenant_id
        - name

    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        resource_type:
          type: string
        action:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - tenant_id
        - name
        - resource_type
        - action

    Policy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        version:
          type: integer
        priority:
          type: integer
        condition:
          type: object
        effect:
          type: string
          enum: [allow, deny]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - tenant_id
        - name
        - version
        - priority
        - condition
        - effect

    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        principal_id:
          type: string
          format: uuid
          nullable: true
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        allowed:
          type: boolean
        reason:
          type: string
        context:
          type: object
        evaluated_at:
          type: string
          format: date-time
      required:
        - id
        - tenant_id
        - action
        - resource_type
        - allowed

    AuthorizationRequest:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        principalId:
          type: string
          format: uuid
        action:
          type: string
        resource:
          type: object
          properties:
            type:
              type: string
            id:
              type: string
            attributes:
              type: object
        principal:
          type: object
        context:
          type: object
      required:
        - tenantId
        - principalId
        - action
        - resource

    AuthorizationResponse:
      type: object
      properties:
        allowed:
          type: boolean
        reason:
          type: string
        explanation:
          type: string
        policy_evaluated:
          type: string
          format: uuid
        failed_conditions:
          type: array
          items:
            type: object
        evaluated_at:
          type: string
          format: date-time
        cache_hit:
          type: boolean
      required:
        - allowed
        - reason

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      required:
        - error

    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        count:
          type: integer
      required:
        - data
        - count

  parameters:
    TenantId:
      name: X-Tenant-ID
      in: header
      description: Tenant identifier
      required: true
      schema:
        type: string
        format: uuid

    Limit:
      name: limit
      in: query
      description: Number of items to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100

    Offset:
      name: offset
      in: query
      description: Number of items to skip
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0

paths:
  /authorize:
    post:
      summary: Evaluate authorization decision
      description: Evaluate whether a principal is authorized to perform an action on a resource
      operationId: authorize
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizationRequest'
      responses:
        '200':
          description: Authorization decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health:
    get:
      summary: Health check
      description: Check service health status
      operationId: health
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time
                  service:
                    type: string

  /api/v1/tenants:
    get:
      summary: List tenants
      description: Get all tenants
      operationId: listTenants
      tags:
        - Tenants
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/tenants/{id}:
    get:
      summary: Get tenant
      description: Get a specific tenant by ID
      operationId: getTenant
      tags:
        - Tenants
      parameters:
        - name: id
          in: path
          description: Tenant ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/roles:
    get:
      summary: List roles
      description: Get all roles for the tenant
      operationId: listRoles
      tags:
        - Roles
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '400':
          description: Missing tenant ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/permissions:
    get:
      summary: List permissions
      description: Get all permissions for the tenant
      operationId: listPermissions
      tags:
        - Permissions
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: List of permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '400':
          description: Missing tenant ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/policies:
    get:
      summary: List policies
      description: Get all policies for the tenant
      operationId: listPolicies
      tags:
        - Policies
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '400':
          description: Missing tenant ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/audit:
    get:
      summary: List audit logs
      description: Get audit logs for the tenant
      operationId: listAuditLogs
      tags:
        - Audit
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '400':
          description: Missing tenant ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/logs:
    get:
      summary: List audit logs
      description: Get audit logs for the tenant (Audit Service)
      operationId: listAuditLogsAudit
      tags:
        - Audit
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '400':
          description: Missing tenant ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/export:
    get:
      summary: Export audit logs
      description: Export audit logs as CSV
      operationId: exportAuditLogs
      tags:
        - Audit
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: CSV export of audit logs
          content:
            text/csv:
              schema:
                type: string
        '400':
          description: Missing tenant ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/compliance:
    get:
      summary: Get compliance summary
      description: Get compliance statistics
      operationId: getCompliance
      tags:
        - Audit
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Compliance summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_logs:
                    type: integer
                  denied_count:
                    type: integer
                  allowed_count:
                    type: integer
                  compliance_rate:
                    type: string
        '400':
          description: Missing tenant ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

tags:
  - name: Authorization
    description: Authorization decision endpoints
  - name: Health
    description: Health check endpoints
  - name: Tenants
    description: Tenant management endpoints
  - name: Roles
    description: Role management endpoints
  - name: Permissions
    description: Permission management endpoints
  - name: Policies
    description: Policy management endpoints
  - name: Audit
    description: Audit log endpoints